moved runner state change into runners from theia state change
