+++ backend/app/api/main.py
@@ -1,13 +1,14 @@
 """Main application file for the API."""
 
 from fastapi import APIRouter
-from app.api.routes import auth, users, runners, machines, cloud_connectors, images, app_requests, theia_requests  # Import your new images route
+from app.api.routes import auth, users, runners, machines, cloud_connectors, images, app_requests  # Import your new images route
 
 API_ROOT_PATH: str = '/api' #stripped out of request.url.path by fastAPI or by the proxy? (thinking fastAPI)
 API_VERSION: str = '/v1'#still present in the path, not for docs
 UNSECURE_ROUTES: tuple = (
     f'{API_VERSION}/machine_auth/',
     f'{API_VERSION}/machine_auth',
+    f'{API_VERSION}/runners/[0-9]+/state',
     f'{API_VERSION}/',
     '/docs',
     '/openapi.json',
@@ -22,5 +23,4 @@ api_router.include_router(auth.router, prefix=f"{API_VERSION}/machine_auth", tag
 api_router.include_router(images.router, prefix=f"{API_VERSION}/images", tags=["images"])
 api_router.include_router(machines.router, prefix=f"{API_VERSION}/machines", tags=["machines"])
 api_router.include_router(cloud_connectors.router, prefix=f"{API_VERSION}/cloud_connectors", tags=["cloud_connectors"])
-api_router.include_router(app_requests.router, prefix=f"{API_VERSION}/app_requests", tags=["app_requests"]) # include your new images route
-api_router.include_router(theia_requests.router, prefix=f"{API_VERSION}/theia_requests", tags=["theia_requests"]) # include your new images route
+api_router.include_router(app_requests.router, prefix=f"{API_VERSION}/app_requests", tags=["app_requests"])
