+++ proxy/nginx.conf
@@ -167,11 +167,12 @@ http {
                         })
                         
                         local httpc = http.new()
-                        local res, err = httpc:request_uri("http://backend:8000/v1/theia_requests/update_state", {
-                            method = "POST",
+                        local res, err = httpc:request_uri("http://backend:8000/api/v1/runners/"..jwt_obj.payload.runner_id.."/state", {
+                            method = "PUT",
                             body = payload,
                             headers = {
                                 ["Content-Type"] = "application/json",
+                                ["Access-Token"] = jwt_secret, 
                             }
                         })
